+++ bingads/v10/bulk/entities/__init__.py
@@ -13,6 +13,8 @@ from .bulk_ad_group import *
 from .bulk_keyword import *
 from .bulk_campaign_product_scope import *
 from .bulk_ad_group_product_partition import *
+from .bulk_ad_group_remarketing_list_association import *
+from .bulk_remarketing_list import *
 
 from .ad_extensions import *
 from .bulk_ads import *