+++ bingads/v10/bulk/entities/ad_extensions/__init__.py
@@ -7,4 +7,4 @@ from .bulk_site_links_ad_extensions import *
 from .bulk_app_ad_extensions import *
 from .bulk_callout_ad_extensions import *
 from .bulk_review_ad_extensions import *
-
+from .bulk_structured_snippet_ad_extensions import *